models/requirements.txt
auto_mix_prep==0.2.0
einops==0.7.0
numpy==1.26.2
omegaconf==2.3.0
opt_einsum==3.3.0
pandas==2.1.3
pysam==0.21.0
torchmetrics==1.2.0
tqdm==4.66.1
pyarrow==13.0.0
scikit-learn==1.3.1
transformers==4.29.2
